caracteristicas finalizadas

+++ petfinder-api/petfinder-api/src/main/java/petfinder/petfinderapi/controladores/PetsController.java
@@ -3,15 +3,15 @@ package petfinder.petfinderapi.controladores;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import petfinder.petfinderapi.entidades.Caracteristica;
-import petfinder.petfinderapi.entidades.Pet;
-import petfinder.petfinderapi.entidades.Premio;
+import petfinder.petfinderapi.entidades.*;
 import petfinder.petfinderapi.repositorios.CaracteristicaRepositorio;
 import petfinder.petfinderapi.repositorios.PetRepositorio;
 import petfinder.petfinderapi.repositorios.PremioRepositorio;
 
+import javax.validation.Valid;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/pets")
@@ -26,14 +26,29 @@ public class PetsController {
     @Autowired
     private CaracteristicaRepositorio repositoryCaracteristica;
 
+
+
     public static List<Pet> pets = new ArrayList<>();
     public static List<Premio> premios = new ArrayList<>();
 
     @PostMapping("/post-pet")
-    public ResponseEntity postPet(
-            @RequestBody Pet novoPet) {
-        repositoryPet.save(novoPet);
-        return ResponseEntity.status(201).build();
+    public ResponseEntity<Object> postPet(
+            @RequestBody @Valid Pet novoPet) {
+        if(Objects.nonNull(novoPet)) {
+            repositoryPet.save(novoPet);
+            return ResponseEntity.status(201).build();
+        }
+        return ResponseEntity.status(400).build();
+    }
+
+    @PutMapping("/atualizar-pet/{indice}")
+    public ResponseEntity putPet(@RequestBody Pet petAtualizado, @PathVariable int indice) {
+        if (repositoryPet.existsById(indice)){
+            petAtualizado.setId(indice);
+            repositoryPet.save(petAtualizado);
+            return ResponseEntity.status(200).build();
+        }
+        return ResponseEntity.status(400).build();
     }
 
     @GetMapping("/get-pets")
@@ -49,20 +64,29 @@ public class PetsController {
 
     @GetMapping("/filtro-id-pet/{id}")
     ResponseEntity getByid(@PathVariable int id) {
-        List<Pet> lista = repositoryPet.findById(id);
-        return ResponseEntity.status(200).body(lista);
+        if (repositoryPet.existsById(id)) {
+            List<Pet> lista = repositoryPet.findById(id);
+            return ResponseEntity.status(200).body(lista);
+        }
+        return ResponseEntity.status(404).build();
     }
 
-    @GetMapping("filtro-id-pets/instituicao/{id}")
+    @GetMapping("/filtro-id-pets/instituicao/{id}")
     ResponseEntity getByInstiuicaoId(@PathVariable int id) {
-        List<Pet> lista = repositoryPet.findByFkInstituicao(id);
-        return ResponseEntity.status(200).body(lista);
+        if (repositoryPet.existsById(id)) {
+            List<Pet> lista = repositoryPet.findByFkInstituicao(id);
+            return ResponseEntity.status(200).body(lista);
+        }
+        return ResponseEntity.status(404).build();
     }
 
-    @GetMapping("delete-id-pet/{id}")
+    @DeleteMapping("/delete-id-pet/{id}")
     ResponseEntity deleteByIdpet(@PathVariable int id) {
-        repositoryPet.deleteById(id);
-        return ResponseEntity.status(200).build();
+        if (repositoryPet.existsById(id)) {
+            repositoryPet.deleteById(id);
+            return ResponseEntity.status(200).build();
+        }
+        return ResponseEntity.status(404).build();
     }
 
 
@@ -70,14 +94,28 @@ public class PetsController {
     // Premios
     // ================================================= //
 
-    @PostMapping("post-premio")
+    @PostMapping("/post-premio")
     public ResponseEntity postPremio(
             @RequestBody Premio novoPremio) {
-        repositoryPremio.save(novoPremio);
-        return ResponseEntity.status(201).build();
+
+        if (Objects.nonNull(novoPremio)) {
+            repositoryPremio.save(novoPremio);
+            return ResponseEntity.status(201).build();
+        }
+        return ResponseEntity.status(400).build();
     }
 
-    @GetMapping("get-premios")
+    @PutMapping("/atualizar-premio/{indice}")
+    public ResponseEntity putPremio(@RequestBody Premio premioAtualizado, @PathVariable int indice) {
+        if (repositoryPremio.existsById(indice)){
+            premioAtualizado.setId(indice);
+            repositoryPremio.save(premioAtualizado);
+            return ResponseEntity.status(200).build();
+        }
+        return ResponseEntity.status(400).build();
+    }
+
+    @GetMapping("/get-premios")
     public ResponseEntity getPremios() {
         List<Premio> lista = repositoryPremio.findAll();
 
@@ -90,34 +128,47 @@ public class PetsController {
 
     @GetMapping("/filtro-id-premio/{id}")
     ResponseEntity getByidPremios(@PathVariable int id) {
-        List<Premio> lista = repositoryPremio.findById(id);
-        return ResponseEntity.status(200).body(lista);
+        if (repositoryPremio.existsById(id)) {
+            List<Premio> lista = repositoryPremio.findById(id);
+            return ResponseEntity.status(200).body(lista);
+        }
+        return ResponseEntity.status(404).build();
     }
 
-    @GetMapping("filtro-id/pet/{id}")
-    ResponseEntity getByPetId(@PathVariable int id) {
-        List<Premio> lista = repositoryPremio.findByFkPet(id);
-        return ResponseEntity.status(200).body(lista);
-    }
-
-    @GetMapping("delete-id-premio/{id}")
+    @DeleteMapping("/delete-id-premio/{id}")
     ResponseEntity deleteByIdPremio(@PathVariable int id) {
-        repositoryPremio.deleteById(id);
-        return ResponseEntity.status(200).build();
+        if (repositoryPremio.existsById(id)) {
+            repositoryPremio.deleteById(id);
+            return ResponseEntity.status(200).build();
+        }
+        return ResponseEntity.status(404).build();
     }
 
     // ================================================= //
     // Caracteristicas
     // ================================================= //
 
-    @PostMapping("post-caracteristica")
+    @PostMapping("/post-caracteristica")
     public ResponseEntity postCaracteristica(
             @RequestBody Caracteristica novaCaracteristica) {
-        repositoryCaracteristica.save(novaCaracteristica);
-        return ResponseEntity.status(201).build();
+        if (Objects.nonNull(novaCaracteristica)){
+            repositoryCaracteristica.save(novaCaracteristica);
+            return ResponseEntity.status(201).build();
+        }
+        return ResponseEntity.status(400).build();
     }
 
-    @GetMapping("get-caracteristicas")
+    @PutMapping("/atualizar-caracteristica/{indice}")
+    public ResponseEntity putCaracteristica(@RequestBody Caracteristica caracteristicaAtualizada, @PathVariable int indice) {
+        if (repositoryCaracteristica.existsById(indice)){
+            caracteristicaAtualizada.setId(indice);
+            repositoryCaracteristica.save(caracteristicaAtualizada);
+            return ResponseEntity.status(200).build();
+        }
+        return ResponseEntity.status(400).build();
+    }
+
+    @GetMapping("/get-caracteristicas")
     public ResponseEntity getCaracteristica() {
         List<Caracteristica> lista = repositoryCaracteristica.findAll();
 
@@ -129,14 +180,27 @@ public class PetsController {
     }
 
     @GetMapping("/filtro-id-caracteristica/{id}")
-    ResponseEntity getByIdCaracteristica(@PathVariable int id) {
-        List<Caracteristica> lista = repositoryCaracteristica.findById(id);
-        return ResponseEntity.status(200).body(lista);
+    ResponseEntity getByidCaracteristca(@PathVariable int id) {
+        if (repositoryCaracteristica.existsById(id)) {
+            List<Caracteristica> lista = repositoryCaracteristica.findById(id);
+            return ResponseEntity.status(200).body(lista);
+        }
+        return ResponseEntity.status(404).build();
     }
 
-    @GetMapping("delete-id-caracteristica/{id}")
-    ResponseEntity deleteByIdCaracteristca(@PathVariable int id) {
-        repositoryCaracteristica.deleteById(id);
-        return ResponseEntity.status(200).build();
+
+    @DeleteMapping("/delete-id-caracteristica/{id}")
+    ResponseEntity deleteByIdCaracteristica(@PathVariable int id) {
+        if (repositoryCaracteristica.existsById(id)) {
+            repositoryCaracteristica.deleteById(id);
+            return ResponseEntity.status(200).build();
+        }
+        return ResponseEntity.status(404).build();
     }
+
+    // ================================================= //
+    // Pet Has Caracteristicas
+    // ================================================= //
+
+
 }
